requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

asgiref==3.7.2; python_version >= '3.7'

colorama==0.4.6; platform_system == 'Windows'

django-froala-editor==4.1.0

django-jazzmin==2.6.0

django-multiselectfield==0.1.12

django-tinymce4-lite==1.8.0

django==4.2.3

jsmin==3.0.1

pillow==10.0.0

pypng==0.20220715.0

qrcode[pil]==7.4.2

reportlab==4.0.4

sqlparse==0.4.4; python_version >= '3.5'

typing-extensions==4.7.1; python_version >= '3.7'

tzdata==2023.3; sys_platform == 'win32'

whitenoise==6.5.0

